Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -3,12 +3,18 @@ package com.educandoweb.course.services;
 import java.util.List;
 import java.util.Optional;
 
+import javax.swing.DefaultRowSorter;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
+import com.sun.org.apache.bcel.internal.generic.IF_ACMPEQ;
 
 @Service
 public class UserService {
@@ -29,8 +35,20 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);
+	public void delete(Long id) { 
+		
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+		
+		try {
+			repository.deleteById(id);;
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+		
 	}
 	
 	public User update(Long id, User obj) {
